pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telegram-harness"
version = "0.1.0"
description = "Extensible Telegram bot harness for remote command execution"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "typer>=0.12",
    "httpx>=0.27",
]

[project.scripts]
telegram-harness = "telegram_harness.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
